djangoapp/requirements.txt
Django==5.0
psycopg2-binary==2.9.9
Pillow==10.1.0
django-summernote==0.8.20.0
python-dotenv==1.0.0
django-axes==6.2.0